add animation to themeButton
👌🏽 updated theme to include cursor pointer
😅 changed name to ThemeButton
👌🏽 refactor themeSwitch
👌🏽 add animation to themeSwitch
🐞🐜 need to workout proper opacity rules...
