+++ src/components/ThemeButton/ThemeButton.module.css
@@ -1,0 +1,44 @@
+.modeWrapper {
+    display: flex;
+        justify-content: center;
+        align-items: flex-start;
+        position: fixed;
+        z-index: 1000;
+        top: 5px;
+        left: 30%;
+}
+.modeContainer {
+    position: relative;
+margin: 15px 0 0 10px;
+--rotation: 0;
+width: 200px;
+height: 200px;
+border: 1px solid rgba(0, 0, 0, 0.1);
+border-radius: 50%;
+transform: rotate(0deg);
+transform: rotate(calc(var(--rotation)*1deg));
+transition: transform var(--transition-time) linear;
+}
+
+.modeIconContainer {
+    position: relative;
+    width: 40px;
+    height: 40px;
+    transform: rotate(0deg);
+    transition: transform var(--transition-time) linear, opacity var(--transition-time) var(--transition-time-function);
+    top: -20px; /* -child size/2 */
+    left: 90px; /* parent size/2 - child size/2 */
+}
+.modeIcons{
+    position: relative;
+    transition: opacity var(--transition-time) var(--transition-time-function)
+}
+.lightModeIcon {
+    opacity: 0;
+    transform: rotate(0deg);
+}
+
+.darkModeIcon {
+    opacity: 1;
+    transform: rotate(180deg);
+}
